+++ Rubberduck.VBEEditor/Rubberduck.VBEditor.csproj
@@ -128,8 +128,9 @@
     <Reference Include="System.Windows.Forms" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ComManagement\TypeLibs\TypeInfos.cs" />
     <Compile Include="ComManagement\TypeLibs\TypeLibs.cs" />
-    <Compile Include="ComManagement\TypeLibs\TypeLibsAbstract.cs" />
+    <Compile Include="ComManagement\TypeLibs\TypeLibsSupport.cs" />
     <Compile Include="ComManagement\TypeLibs\TypeLibsAPI.cs" />
     <Compile Include="ComManagement\ReferenceEqualityComparer.cs" />
     <Compile Include="ComManagement\ComSafe.cs" />
